Finish DebugInfo ISerializable
